Final functions to make it work with static inputs

+++ structs/local_items_struct.go
@@ -6,10 +6,10 @@ type Format struct {
 }
 
 type LocalItems = map[string]struct {
-	Name   string  `json:"name"`
-	Type   string  `json:"type"`
-	Min    Format  `json:"min"`
-	Max    *Format `json:"max,omitempty"`
-	Onhit  bool    `json:"onhit,omitempty"`
-	Effect []int   `json:"effect,omitempty"`
+	Name   string    `json:"name"`
+	Type   string    `json:"type"`
+	Min    Format    `json:"min"`
+	Max    *Format   `json:"max,omitempty"`
+	Onhit  bool      `json:"onhit,omitempty"`
+	Effect []float64 `json:"effect,omitempty"`
 }
